mergeBreakpoints: Breaks down the folder structure
